+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Remove usage of legacy `childContextTypes` API
 - Refactor Field components to use arrow function syntax
 - Replace react-autocomplete with Downshift in the autocomplete component
+- Add LocationIQ as supported map provider with access token field and gallery style
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes